api: touch api skeleton
Introduces a skeleton for the Houdini API. This exposes a Unix socket (path is
configurable, /var/run/houdini.sock by default) that is bound to an HTTP server. At the
moment, the only available route is `/ping` which replies with "pong". The plan is to
implement actual routes as a follow-up to enable users to run exploits using a client
application. This will enable us to run a Houdini "client" on the host and "server" in
a virtual machine. The result will be the ability to run Houdini exploits in fully
configurable environments.

Example Usage:

    # In one terminal:
    $ cargo run -- api serve

    # In another terminal:
    $ sudo curl --unix socket /var/run/houdini.sock http://localhost/ping

Signed-off-by: William Findlay <will@isovalent.com>