api: add a run trick endpoint
Add a new POST endpoint `/trick` that will the Trick specified as a JSON payload. Returns
a JSON serialized TrickReport or a status code on error.

N.B. For now the body of the handler is essentially a NOP. I'll be adding the
functionality to run the tricks in a subsequent patch.

Example Usage:

    # In one terminal:
    $ cargo run -- api serve

    # In another terminal:
    $ sudo curl --unix-socket /var/run/houdini.sock \
        -H "Content-Type: application/json" \
        -X POST http://localhost/trick \
        -d '{"name": "foo", "steps": []}'

Signed-off-by: William Findlay <will@isovalent.com>